Work on migration

+++ src/main/java/com/lenis0012/bukkit/loginsecurity/modules/migration/LegacyMigration.java
@@ -1,6 +1,9 @@
 package com.lenis0012.bukkit.loginsecurity.modules.migration;
 
+import com.avaje.ebean.SqlQuery;
+import com.avaje.ebean.SqlUpdate;
 import com.lenis0012.bukkit.loginsecurity.LoginSecurity;
+import com.lenis0012.bukkit.loginsecurity.modules.storage.StorageModule;
 import com.lenis0012.pluginutils.PluginHolder;
 
 import java.io.*;
@@ -16,30 +19,44 @@ public class LegacyMigration extends Migration {
     @Override
     public boolean executeAutomatically() {
         final PluginHolder plugin = LoginSecurity.getInstance();
-        final File file = new File(plugin.getDataFolder(), "users.db");
-        return file.exists();
+        final StorageModule storage = plugin.getModule(StorageModule.class);
+        if(storage.isRunningMySQL()) {
+            try {
+                SqlQuery query = plugin.getDatabase().createSqlQuery("SELECT COUNT(*) FROM users");
+                int rows = (Integer) query.findUnique().values().toArray(new Object[0])[0];
+                return rows > 0;
+            } catch(Exception e) {
+                e.printStackTrace(); //TODO: Remove debug
+                return false;
+            }
+        } else {
+            final File file = new File(plugin.getDataFolder(), "users.db");
+            return file.exists();
+        }
     }
 
     @Override
-    public boolean canExecute() {
+    public boolean canExecute(String[] params) {
         return false;
     }
 
     @Override
-    public boolean execute() {
+    public boolean execute(String[] params) {
         final PluginHolder plugin = LoginSecurity.getInstance();
+        final StorageModule storage = plugin.getModule(StorageModule.class);
         final File file = new File(plugin.getDataFolder(), "users.db");
         final Logger logger = plugin.getLogger();
+        final String platform = storage.isRunningMySQL() ? "mysql" : "sqlite";
 
-        final String url = "jdbc:sqlite:" + file.getPath();
-        final String driver = "org.sqlite.JDBC";
+        String url = "jdbc:sqlite:" + file.getPath();
+        String driver = "org.sqlite.JDBC";
 
         // Load driver
         log("Loading driver...");
         try {
             Class.forName(driver);
         } catch(Exception e) {
-            logger.log(Level.SEVERE, "Failed to load SQLite driver.");
+            logger.log(Level.SEVERE, "Failed to load SQL driver.");
             return false;
         }
 
@@ -48,7 +65,7 @@ public class LegacyMigration extends Migration {
         final StringBuilder builder = new StringBuilder();
         BufferedReader reader = null;
         try {
-            final InputStream input = plugin.getResource("sql/sqlite/legacy_upgrade.sql");
+            final InputStream input = plugin.getResource("sql/" + platform + "/legacy_upgrade.sql");
             reader = new BufferedReader(new InputStreamReader(input));
             String line;
             while((line = reader.readLine()) != null) {
@@ -65,39 +82,62 @@ public class LegacyMigration extends Migration {
             }
         }
 
-        log("Creating backup...");
-        try {
-            copyFile(file, new File(plugin.getDataFolder(), "users.backup.db"));
-        } catch(IOException e) {
-            logger.log(Level.SEVERE, "Failed to create backup", e);
+        // MySQL: Apply upgrade script
+        if(storage.isRunningMySQL()) {
+            SqlUpdate update = plugin.getDatabase().createSqlUpdate(builder.toString());
+            update.execute();
         }
 
-        log("Executing query, this might take a while...");
-        Connection connection = null;
-        try {
-            connection = DriverManager.getConnection(url);
-            Statement statement = connection.createStatement();
-//            statement.setQueryTimeout(120);
-            statement.execute(builder.toString());
-        } catch(SQLException e) {
-            logger.log(Level.SEVERE, "Failed to run SQL query", e);
-            return false;
-        } finally {
-            if(connection != null) {
-                try {
-                    connection.close();
-                } catch(SQLException e) {}
+        // SQLite: Apply upgrade script
+        if(!storage.isRunningMySQL()) {
+            // Create backup
+            log("Creating backup...");
+            try {
+                copyFile(file, new File(plugin.getDataFolder(), "users.backup.db"));
+            } catch(IOException e) {
+                logger.log(Level.SEVERE, "Failed to create backup", e);
             }
-        }
 
-        log("Finalizing...");
-        File newFile = new File(plugin.getDataFolder(), "LoginSecurity.sql");
-        if(newFile.exists()) {
-            log("Deleting old database...");
+            // Execute upgrade query
+            log("Executing query, this might take a while...");
+            Connection connection = null;
+            try {
+                connection = DriverManager.getConnection(url);
+                Statement statement = connection.createStatement();
+                statement.execute(builder.toString());
+            } catch(SQLException e) {
+                logger.log(Level.SEVERE, "Failed to run SQL query", e);
+                return false;
+            } finally {
+                if(connection != null) {
+                    try {
+                        connection.close();
+                    } catch(SQLException e) {}
+                }
+            }
+
+            // Update files
+            log("Finalizing...");
+            File newFile = new File(plugin.getDataFolder(), "LoginSecurity.sql");
+            if(newFile.exists()) {
+                log("Deleting old database...");
+                long startTime = System.currentTimeMillis();
+                while(!newFile.delete()) {
+                    if(startTime + 30000L > System.currentTimeMillis()) {
+                        logger.log(Level.WARNING, "Failed to delete old database, please follow instructions on wiki to resolve!");
+                        return false;
+                    }
+                    try {
+                        Thread.sleep(50L);
+                    } catch(InterruptedException e) {
+                    }
+                }
+            }
+            log("Renaming new database...");
             long startTime = System.currentTimeMillis();
-            while(!newFile.delete()) {
+            while(!file.renameTo(newFile)) {
                 if(startTime + 30000L > System.currentTimeMillis()) {
-                    logger.log(Level.WARNING, "Failed to delete old database, please follow instructions on wiki to resolve!");
+                    logger.log(Level.WARNING, "Failed to rename new database, please follow instructions on wiki to resolve!");
                     return false;
                 }
                 try {
@@ -106,24 +146,18 @@ public class LegacyMigration extends Migration {
                 }
             }
         }
-        log("Renaming new database...");
-        long startTime = System.currentTimeMillis();
-        while(!file.renameTo(newFile)) {
-            if(startTime + 30000L > System.currentTimeMillis()) {
-                logger.log(Level.WARNING, "Failed to rename new database, please follow instructions on wiki to resolve!");
-                return false;
-            }
-            try {
-                Thread.sleep(50L);
-            } catch(InterruptedException e) {
-            }
-        }
 
+        storage.applyMissingUpgrades();
         return true;
     }
 
     @Override
     public String getName() {
         return "Legacy";
+    }
+
+    @Override
+    public int minParams() {
+        return 0; //
     }
 }
